orm_search/app/requirements.txt
Flask==2.2.2
flask_sqlalchemy==3.0.0
psycopg2-binary==2.9.3
python-dotenv==0.21.0